Put validators in seperate file, write tests for them
